extract receipt and amount utils
